Introduce backend-defaults package
