+++ packages/backend-app-api/src/services/implementations/urlReaderService.ts
@@ -23,6 +23,7 @@ import {
   urlReaderServiceRef,
 } from '@backstage/backend-plugin-api';
 
+/** @public */
 export const urlReaderFactory = createServiceFactory({
   service: urlReaderServiceRef,
   deps: {
