+++ packages/backend-defaults/package.json
@@ -1,0 +1,47 @@
+{
+  "name": "@backstage/backend-defaults",
+  "description": "Backend defaults used by Backstage backend apps",
+  "version": "0.0.0",
+  "main": "src/index.ts",
+  "types": "src/index.ts",
+  "private": false,
+  "publishConfig": {
+    "access": "public",
+    "main": "dist/index.cjs.js",
+    "types": "dist/index.d.ts",
+    "alphaTypes": "dist/index.alpha.d.ts"
+  },
+  "backstage": {
+    "role": "node-library"
+  },
+  "homepage": "https://backstage.io",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/backstage/backstage",
+    "directory": "packages/backend-defaults"
+  },
+  "keywords": [
+    "backstage"
+  ],
+  "license": "Apache-2.0",
+  "scripts": {
+    "build": "backstage-cli package build --experimental-type-build",
+    "lint": "backstage-cli package lint",
+    "test": "backstage-cli package test",
+    "prepack": "backstage-cli package prepack",
+    "postpack": "backstage-cli package postpack",
+    "clean": "backstage-cli package clean",
+    "start": "backstage-cli package start"
+  },
+  "dependencies": {
+    "@backstage/backend-app-api": "^0.1.1-next.0",
+    "@backstage/backend-plugin-api": "^0.1.1-next.0"
+  },
+  "devDependencies": {
+    "@backstage/cli": "^0.18.1-next.0"
+  },
+  "files": [
+    "dist",
+    "alpha"
+  ]
+}
