+++ packages/backend-defaults/src/index.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 The Backstage Authors
+ * Copyright 2022 The Backstage Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,11 +14,5 @@
  * limitations under the License.
  */
 
-/**
- * Core API used by Backstage backend apps.
- *
- * @packageDocumentation
- */
-
-export * from './wiring';
-export * from './services/implementations';
+export type { CreateBackendOptions } from './types';
+export { createBackend } from './CreateBackend';
