+++ packages/backend-defaults/src/types.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 The Backstage Authors
+ * Copyright 2022 The Backstage Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,11 +14,11 @@
  * limitations under the License.
  */
 
-/**
- * Core API used by Backstage backend apps.
- *
- * @packageDocumentation
- */
+import { AnyServiceFactory } from '@backstage/backend-plugin-api';
 
-export * from './wiring';
-export * from './services/implementations';
+/**
+ * @public
+ */
+export interface CreateBackendOptions {
+  serviceFactories?: AnyServiceFactory[];
+}
