+++ packages/backend-next/package.json
@@ -25,7 +25,7 @@
     "clean": "backstage-cli package clean"
   },
   "dependencies": {
-    "@backstage/backend-app-api": "^0.1.1-next.0",
+    "@backstage/backend-defaults": "^0.0.0",
     "@backstage/plugin-catalog-backend": "^1.3.1-next.0",
     "@backstage/plugin-scaffolder-backend": "^1.5.0-next.0"
   },
